requirements.txt
Django==1.11.9
django-pyodbc==1.1.3
django-pyodbc-azure==1.11.0.0
djangorestframework==3.7.7
numpy==1.14.0
pip==9.0.1
pyodbc==4.0.21
pytz==2017.3
scipy==1.0.0
setuptools==28.8.0
